+++ src/br/gov/caixa/app/Models/Contas/ContaInvestimento/AbreContaInvestimento.java
@@ -9,6 +9,7 @@ import br.gov.caixa.app.Models.Services.HistoricoAcoes.ListaAcoes;
 import br.gov.caixa.app.Models.Services.HistoricoAcoes.RegistraAcoes;
 import br.gov.caixa.app.Models.Users.Cliente;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public abstract class AbreContaInvestimento implements CriaContaInvestimento {
@@ -16,8 +17,8 @@ public abstract class AbreContaInvestimento implements CriaContaInvestimento {
         public static ContaInvestimento criaContaInvestimento(Date dataDeAtualizacao, Status status, Cliente cliente, String IdUsuario) {
         int geraId = (int) Math.random();
         String id = Integer.toString(geraId);
-        float saldo = 0;
-        HistoricoAcoes historicoAcoes = new HistoricoAcoes(dataDeAtualizacao, "Abertura da conta Investimento", 0, 0, null, null);
+        BigDecimal saldo = BigDecimal.valueOf(0);
+        HistoricoAcoes historicoAcoes = new HistoricoAcoes(dataDeAtualizacao, "Abertura da conta Investimento", BigDecimal.valueOf(0), BigDecimal.valueOf(0), null, null);
         ListaAcoes listaAcoesContaInvestimento = new ListaAcoes();
         RegistraAcoes.registraAcoes(listaAcoesContaInvestimento, historicoAcoes);
         return new ContaInvestimento(id, saldo, listaAcoesContaInvestimento, dataDeAtualizacao, status, cliente, IdUsuario);
