+++ src/br/gov/caixa/app/Models/Contas/ContaInvestimento/ContaInvestimento.java
@@ -6,6 +6,7 @@ import br.gov.caixa.app.Models.Services.HistoricoAcoes.HistoricoAcoes;
 import br.gov.caixa.app.Models.Services.HistoricoAcoes.ListaAcoes;
 import br.gov.caixa.app.Models.Users.Cliente;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class ContaInvestimento extends Conta {
@@ -20,7 +21,7 @@ public class ContaInvestimento extends Conta {
         this.dataInvestmento = dataInvestmento;
     }
 
-    public ContaInvestimento(String id, float saldo, ListaAcoes listaAcoes, Date dataDeAtualizacao, Status status, Cliente cliente, String IdUsuario) {
+    public ContaInvestimento(String id, BigDecimal saldo, ListaAcoes listaAcoes, Date dataDeAtualizacao, Status status, Cliente cliente, String IdUsuario) {
         super(id, saldo, listaAcoes, dataDeAtualizacao, status, cliente);
     }
 }
