+++ src/br/gov/caixa/app/Models/Services/HistoricoAcoes/HistoricoAcoes.java
@@ -2,19 +2,18 @@ package br.gov.caixa.app.Models.Services.HistoricoAcoes;
 
 import br.gov.caixa.app.Models.Users.Cliente;
 
-import java.util.ArrayList;
+import java.math.BigDecimal;
 import java.util.Date;
-import java.util.List;
 
 public class HistoricoAcoes {
     private Date data;
     private String identficacao;
-    private double valorPretendito;
-    private double valorReal;
+    private BigDecimal valorPretendito;
+    private BigDecimal valorReal;
     private Cliente usuarioOrigem;
     private Cliente usuarioDestino;
 
-    public HistoricoAcoes(Date data, String identficacao, double valorPretendito, double valorReal, Cliente usuarioOrigem, Cliente usuarioDestino) {
+    public HistoricoAcoes(Date data, String identficacao, BigDecimal valorPretendito, BigDecimal valorReal, Cliente usuarioOrigem, Cliente usuarioDestino) {
         this.data = data;
         this.identficacao = identficacao;
         this.valorPretendito = valorPretendito;
@@ -31,11 +30,11 @@ public class HistoricoAcoes {
         return identficacao;
     }
 
-    public double getValorPretendito() {
+    public BigDecimal getValorPretendito() {
         return valorPretendito;
     }
 
-    public double getValorReal() {
+    public BigDecimal getValorReal() {
         return valorReal;
     }
