+++ src/br/gov/caixa/app/Models/Services/Saque/ConsultarSaldo.java
@@ -2,8 +2,10 @@ package br.gov.caixa.app.Models.Services.Saque;
 
 import br.gov.caixa.app.Models.Common.Conta;
 
+import java.math.BigDecimal;
+
 public class ConsultarSaldo {
-    public static float consultaSaldo(Conta conta){
+    public static BigDecimal consultaSaldo(Conta conta){
 
         return conta.getSaldo();
     }
